pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "remote-status-notifier"
version = "0.1.0"
description = "Real-time remote desktop user detection and Slack notification agent."
authors = [{ name = "wyatthoho", email = "wyatthoho@gmail.com" }]
dependencies = [
    "psutil == 6.1.1",
    "slack_sdk == 3.34.0",
    "python-dotenv >= 1.2.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = { file = "LICENSE" }

[project.scripts]
notifier = "notifier.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = ["pytest >= 8.3.4"]
